+++ runtime/regiontemplates/comparativeanalysis/pixelcompare/PixelCompare.cpp
@@ -4,7 +4,7 @@
 
 #include "PixelCompare.h"
 
-PixelCompare::PixelCompare(DenseDataRegion2D *dr1, DenseDataRegion2D *dr2, int *diffPixels) {
+PixelCompare::PixelCompare(DenseDataRegion2D *dr1, DenseDataRegion2D *dr2, float *diffPixels) {
     this->dr1 = dr1;
     this->dr2 = dr2;
     this->diff = diffPixels;
@@ -12,7 +12,7 @@ PixelCompare::PixelCompare(DenseDataRegion2D *dr1, DenseDataRegion2D *dr2, int *
 
 bool PixelCompare::run(int procType, int tid) {
     uint64_t t1 = Util::ClockGetTimeProfile();
-    int compId = diff[0];
+    float compId = diff[0];
     cv::Mat image1 = this->dr1->getData();
     cv::Mat image2 = this->dr2->getData();
     /*std::cout << "dr1->id: "<< dr1->getId() << " version: " << dr1->getVersion() << std::endl;
@@ -27,8 +27,8 @@ bool PixelCompare::run(int procType, int tid) {
     int nonZeroRef = countNonZero(image2);
 
     std::cout << "diffPixels: " << diffPixels << " nonZeroRef: " << nonZeroRef << std::endl;
-    this->diff[0] = diffPixels;
-    this->diff[1] = nonZeroRef;
+    this->diff[0] = (float) diffPixels;
+    this->diff[1] = (float) nonZeroRef;
     std::cout << "CompId: " << compId << " diffPixels: " << this->diff[0] << " reference: " << this->diff[1] <<
     std::endl;
